[Frontend-GNOME] Extract the icon download and caching logic
Instead of having it in the ProtocolChatView, extract the logic to an
IconCache class which we can use for more uses.

Currently it is still very coupled to trying to download the favicon
from a site, but this is a first step in making it reusable.
