+++ src/Frontend-GNOME/Makefile.am
@@ -126,6 +126,7 @@ FILES = \
 	Frontend.cs \
 	GnomeUI.cs \
 	Gtk.WindowExtensions.cs \
+	IconCache.cs \
 	IndicateManager.cs \
 	Main.cs \
 	MainWindow.cs \
